# Optimize GeoArrow/Arrow IO