--- a/.gitattributes
+++ b/.gitattributes
@@ -1,1 +1,3 @@
 shapely/_version.py export-subst
+src/geoarrow.h linguist-vendored
+src/geoarrow.c linguist-vendored