--- a/shapely/_geoarrow.pyx
+++ b/shapely/_geoarrow.pyx
@@ -264,15 +264,14 @@ cdef class ArrayBuilder:
         cdef GEOSGeometry* input_chunk[1024]
         cdef int rc
         cdef size_t n_appended = 0
-        n_appended_total = 0
-        cdef size_t n_full_chunks_in = len(geometries) // 1024
+        cdef int64_t n_appended_total = 0
+        cdef int64_t i = 0
 
-        for chunk_in_i in range(n_full_chunks_in):
-            i_begin = (chunk_in_i * 1024)
-            i_end = i_begin + 1024
-            for geom, chunk_i in zip(geometries[i_begin:i_end], range(1024)):
-                if not PyGEOS_GetGEOSGeometry(<PyObject*>geom, &(input_chunk[chunk_i])):
-                    raise RuntimeError(f"PyGEOS_GetGEOSGeometry() failed at chunk {chunk_in_i}[{chunk_i}]")
+        for chunk_in_i in range(len(geometries) // 1024):
+            for chunk_i in range(1024):
+                if not PyGEOS_GetGEOSGeometry(<PyObject*>(geometries[i]), &(input_chunk[chunk_i])):
+                    raise RuntimeError(f"PyGEOS_GetGEOSGeometry(geometries[{i}])")
+                i += 1
 
             with nogil:
                 rc = GeoArrowGEOSArrayBuilderAppend(self._ptr, input_chunk, 1024, &n_appended)
@@ -282,20 +281,17 @@ cdef class ArrayBuilder:
 
             n_appended_total += n_appended
 
-        # Last chunk
-        i_begin = n_full_chunks_in * 1024
-        i_end = len(geometries)
-        cdef size_t n_remaining = i_end - i_begin
+        cdef int64_t n_remaining = len(geometries) % 1024
+        for chunk_i in range(n_remaining):
+            if not PyGEOS_GetGEOSGeometry(<PyObject*>(geometries[i]), &(input_chunk[chunk_i])):
+                raise RuntimeError(f"PyGEOS_GetGEOSGeometry() failed at last chunk[{chunk_i}]")
+            i += 1
 
-        if n_remaining > 0:
-            for geom, chunk_i in zip(geometries[i_begin:i_end], range(n_remaining)):
-                if not PyGEOS_GetGEOSGeometry(<PyObject*>geom, &(input_chunk[chunk_i])):
-                        raise RuntimeError(f"PyGEOS_GetGEOSGeometry() failed at last chunk[{chunk_i}]")
-            with nogil:
-                    rc = GeoArrowGEOSArrayBuilderAppend(self._ptr, input_chunk, n_remaining, &n_appended)
-            # TODO: check EOVERFLOW, e.g., WKB >2GB reached so we add a chunk and try again
-            if rc != GEOARROW_GEOS_OK:
-                self._raise_last_error(rc, "GeoArrowGEOSArrayBuilderAppend()")
+        with nogil:
+                rc = GeoArrowGEOSArrayBuilderAppend(self._ptr, input_chunk, n_remaining, &n_appended)
+        # TODO: check EOVERFLOW, e.g., WKB >2GB reached so we add a chunk and try again
+        if rc != GEOARROW_GEOS_OK:
+            self._raise_last_error(rc, "GeoArrowGEOSArrayBuilderAppend()")
 
             n_appended_total += n_appended