… pyproject.toml
… shapely/tests/test_geoarrow.py
… src/geoarrow.c
… src/geoarrow.h
… src/geoarrow_geos.c
… src/geoarrow_geos.h
… vendor_geoarrow.sh