@jorisvandenbossche 2023-12-20
Just for clarity of the scope of the custom code _for shapely_: all the .c/.h files are currently vendored (from https://github.com/geoarrow/geoarrow-c-geos and https://github.com/geoarrow/geoarrow-c), and essentially only `_geoarrow.pyx`/`geoarrow.py` is new code here?