@kylebarron 2024-03-18
I think `to_pyarrow` but `from_arrow` might be a little confusing for users. Do you expect a `to_nanoarrow` as well? Should we have `to_arrow` with a potential keyword argument in the future to switch arrow impl?