@kylebarron 2024-03-18
> Good call! I was thinking that`to_arrow()` should return a `GeoArrowExporter`. Then, it's the caller's responsibility to invoke the dependency they had in mind.

That extra level of indirection seems like it would be annoying for most users. I'd be +1 on conditionally importing either pyarrow or nanoarrow and returning those directly.