@kylebarron 2024-03-18
> Something like lonboard might want to instantiate the `GeoArrowExporter` directly to avoid a situation where the user has neither `nanoarrow` nor `pyarrow` installed

Lonboard depends on pyarrow directly. It would be _lovely_ to remove this dependency (especially as that's the biggest blocker to getting lonboard to work in Wasm in jupyterlite/pyodide), but we rely heavily on `pyarrow.Table.from_pandas`, which I don't really want to reimplement myself