@jorisvandenbossche 2024-03-19
> but we rely heavily on `pyarrow.Table.from_pandas`, which I don't really want to reimplement myself in Rust

We should maybe reimplement that in nanoarrow ..