@jorisvandenbossche 2024-03-19
@paleolimbot apologies for the delay in response here. And thanks for exploring this!

Some general concerns / thoughts:

- Shapely is a volunteer-only project with limited maintainer capacity (we are already too slow in reviewing and are not always able to keep up with new GEOS features), and this PR is adding a _lot_ of code. I know part of it is vendored (but I don't know if that code is already used elsewhere as well?), but we still need to understand and know that (not much documented) vendored code well enough to review and maintain the shapely-specific code. I am a bit afraid about the current scope of code addition here.
- Some parts of what is implemented here feels somewhat duplicative with existing functionality of shapely. For example, the type inference could maybe also be done with a shapely alternative based on `np.unique(shapely.get_type_id(arr))` ? While this might mostly be part of the vendored code, from shapely's point of view it might still be duplicative additions
- Further thinking about the scope, the main thing we want here is implement the GEOS-specific aspects about producing/consuming Arrow data (i.e. those parts where you need to build against GEOS). So it might also be an option to focus on that, and leave it at capsules, not bothering with pyarrow or geoarrow-pyarrow or nanoarrow or ..?
- Should we advocate for moving some core functionality of this into GEOS itself? (that might also be useful for GDAL?)
- I did have some preliminary cython code that just optimizes `to_ragged_array` (https://github.com/pygeos/pygeos/pull/93). From a quick test using your water_line test data, that's actually still a bit faster than this PR (but of course not yet that feature-full, like missing values support). Have to profile a bit in more detail to understand this better.