@Oreilles 2024-08-24
Considering that it is possible with your work to call `from_arrow` and parse EWKB data while preserving the SRID in the newly created geometries, do you think it would be possible to also write EWKB (with SRID included) with the function `to_pyarrow` ? It currently isn't the case, meaning a round-trip with these two functions would result in a loss of information.